# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tenant/addon/{featureKey}": {
      "get": {
        "operationId": "entitlements_addonCapability",
        "parameters": [
          {
            "name": "featureKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Entitlements"
        ]
      }
    },
    "/api/v1/platform/tenants/{tenantId}/entitlements": {
      "get": {
        "operationId": "entitlements_listEntitlements",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Entitlements"
        ]
      },
      "post": {
        "operationId": "entitlements_setEntitlement",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Entitlements"
        ]
      }
    },
...
    "/api/v1/tenant/customers/{id}": {
      "get": {
        "operationId": "customers_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "operationId": "customers_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ]
      },
      "delete": {
        "operationId": "customers_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/tenant/customers/{id}/account": {
      "get": {
        "operationId": "customers_account",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ]
      },
      "post": {
        "operationId": "customers_bind",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/tenant/customers/{id}/orders": {
      "get": {
        "operationId": "customers_orders",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Customers"
        ]
      }
    },
...
    "/api/v1/tenant/disputes/{disputeId}": {
      "get": {
        "operationId": "disputes_detail",
        "parameters": [
          {
            "name": "disputeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/tenant/orders/{orderId}/disputes": {
      "get": {
        "operationId": "disputes_list",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Disputes"
        ]
      },
      "post": {
        "operationId": "disputes_open",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Disputes"
        ]
      }
    },
...
    "/api/v1/tenant/disputes/{disputeId}/resolve": {
      "post": {
        "operationId": "disputes_resolve",
        "parameters": [
          {
            "name": "disputeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Disputes"
        ]
      }
    },
...
  },
  "info": {
    "title": "陪玩门店多租户 SaaS API",
    "description": "一期 API。金额字段一律为十进制字符串分（MoneyFen），禁止 number/浮点。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "tenant/platform access token"
      }
    },
...
  }
}